Update Player.java
changed data type of points from int to double

+++ Player.java
@@ -3,7 +3,7 @@ import Team;
 public class Player
 {
     private Team team;
-    private int points;
+    private double points;
     private String name;
     private String position;
     private int weeklyPoints;
